refactor out handle
